put saturation brightness controller to own view
